Add files via upload
This is what I have so far. This is just built to test the 11th column. When ran I get this error "DtypeWarning: Columns (28) have mixed types. Specify dtype option on import or set low_memory=False. data = pd.read_csv('test.csv')". The program still executes and displays a graph. However I am not quite sure it is an accurate representation of the data.